requirements.txt
python-telegram-bot==13.15
google-generativeai>=0.3.2
Pillow>=10.0.0
ccxt>=4.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
